js/xpconnect/tests/mochitest/test_bug564330.html


<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=564330
-->
<head>
  <title>Test for Bug 560351</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=564330">Mozilla Bug 564330</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
function getAllowXML(){
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  Components.utils.import("resource://gre/modules/Services.jsm");
  return Services.prefs.getBoolPref("javascript.options.xml.content");
}

function setAllowXML(value){
  // Set the value for e4x
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  Components.utils.import("resource://gre/modules/Services.jsm");
  Services.prefs.setBoolPref("javascript.options.xml.content", value);
}

// Save the initial preference
var default_allow_xml = getAllowXML();
function restoreAllowXML(){
    setAllowXML(default_allow_xml);
}

setAllowXML(true);
</script>
<script>
/** Test for Bug 564330 **/
var pass = false;
try {
    document.createElement("span").setUserData("x", eval('<y/>'), null);
} catch (e) {
    pass = true;
}

</script>
<script>

ok(pass, "pass was set to true");
restoreAllowXML();
</script>
</pre>
</body>
</html>
